requirements.txt
networkx~=3.4.2
numpy~=1.26.4
scikit-learn~=1.6.1
scipy~=1.12.0
asyncio~=3.4.3
openai~=1.61.0
requests~=2.32.3
mcp~=1.5.0
pydantic~=2.11.1
playwright~=1.51.0
rich~=13.9.4
neo4j~=5.28.1
rdflib~=7.1.4
colorama~=0.4.6
pypdf2~=3.0.1
torch~=2.4.1
jieba3k~=0.35.1
dotenv~=0.9.9
python-dotenv~=1.1.0
transformers~=4.50.3
pipmaster~=0.7.2
pyvis~=0.3.2
sentencepiece~=0.2.0
tenacity~=9.0.0
configparser~=7.2.0
sqlalchemy~=2.0.40
redis~=5.2.1
pymilvus~=2.5.6
asyncpg~=0.30.0
pyjwt~=2.8.0
fastapi~=0.115.12
aiofiles~=23.2.1
pyuca~=1.2
starlette~=0.46.1
uvicorn~=0.34.0
aiohttp~=3.11.14
ollama~=0.4.7
anthropic~=0.49.0
future~=1.0.0
httpx~=0.28.1
setuptools~=75.8.0
jsonlines~=4.0.0
tqdm~=4.67.1
xmltodict~=0.14.2
docx2markdown~=0.1.1
gradio~=3.50.2
datasets~=3.4.1
evaluate~=0.4.3
accelerate~=0.26.1
peft~=0.4.0
regex~=2024.11.6
nltk~=3.9.1
beautifulsoup4~=4.13.3
pandas~=2.2.3
langchain~=0.3.25
matplotlib~=3.10.1
tabulate~=0.9.0
dspy~=2.6.24
pytest~=7.4.4
zhipuai~=2.1.5.20250421
baidusearch~=1.0.3
tiktoken~=0.7.0
docker~=7.1.0
loguru~=0.7.3
boto3~=1.37.23